backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
pydantic>=2.7.0
confluent-kafka>=2.3.0
fastavro>=1.9.4
redis>=5.0.3
python-multipart>=0.0.9
python-dotenv>=1.0.1
python-socketio>=5.11.2
kafka-python>=2.0.2

# Security Dependencies
PyJWT>=2.8.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
